requirements.txt
streamlit
ultralytics
Pillow
requests
geocoder